+++ src/main/java/com/imsweb/validation/internal/IterativeProcessor.java
@@ -1,5 +1,5 @@
-/**
- *
+/*
+ * Copyright (C) 2004 Information Management Services, Inc.
  */
 package com.imsweb.validation.internal;